Extract superclass from Blueprint and Bottling elements
